requirements.txt
Django>=3.2,<3.3
djangorestframework==3.13.1
psycopg2==2.9.3
pandas==2.0.3
openpyxl==3.0.10
requests==2.31.0
pytz==2022.7
pymongo==4.5.0
redis==3.5.3
django-rest-swagger>=2.2.0, <2.3.0
drf-yasg==1.21.5
django-redis==5.0.0